refactor util functions; update tests; bump version
