+++ tdtax/test/test_validate.py
@@ -1,5 +1,12 @@
-import tdtax
+from tdtax import schema
+from tdtax.validate import is_valid
 
 
-def test():
-    tdtax.validate(tdtax.taxonomy, tdtax.taxonomy)
+def test_validate_bad_yaml():
+    ret = is_valid("tdtax/test/bad.yaml", schema)
+    assert ret.find("not against") != -1
+
+
+def test_validate_top_yaml():
+    ret = is_valid("tdtax/top.yaml", schema)
+    assert ret.find("valid") != -1
